opportunityhub/requirements.txt
asgiref==3.9.1
asttokens==3.0.0
boto3==1.38.27
botocore==1.38.27
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cramjam==2.10.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Django==5.2.5
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.0
fastparquet==2024.11.0
fonttools==4.58.0
fsspec==2025.5.1
greenlet==3.2.2
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
PyAthena==3.14.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
s3transfer==0.13.0
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.1.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
